Add files via upload
I think I managed the import and converted my file so it is properply reading the results from Results...py, but unable to test due to lack of access with supabase. I will ask Bam to test tmr and if it doesn't work, I will work with noah to get things running on my end as well. 

Still need to apply coding conventions.