Add files via upload
Nobody needs to use MakeModel, so I left it as a ipynb file, As for how we will read the models and display outputs, I will make a shortened version of the last cell to read off of supabase and clean up the display so it doesn't actually need to display all the extra garbage. Still need to figure out how noah and bam wants the output formatted.