Add files via upload
I got it to run with a single photo on ipynb but I am having trouble with some of the imports ( specifically tensorflow ) when trying to run the python file. If someone can test, please let me know if can get it to give an output. Should be a return of a dict, but if you just need to test, you can uncomment the print( returnStatement ).